requirements.txt
python-docx>=1.1.0
Pillow>=10.0.0
pillow-heif>=0.21.0
pyinstaller>=6.0.0
